Clippy fixes

+++ edgeless_function/src/lcg.rs
@@ -59,6 +59,6 @@ mod test {
         let mut lcg = Lcg::new(42);
         let matrix = random_matrix(&mut lcg, 1000);
         assert_eq!(1000 * 1000, matrix.len());
-        assert_ne!(0.0 as f32, matrix.iter().sum());
+        assert_ne!(0.0_f32, matrix.iter().sum());
     }
 }
